requirements.txt
flask==1.1.1
flask-sqlalchemy==2.4.0
google-cloud-storage==1.16.1
psycopg2-binary==2.8.3
sendgrid==6.0.5
# The cidc_api_modules package
cidc-api-modules==0.1.0
